python/requirements.txt
flask
telethon
pysocks
